pyproject.toml
[tool.poetry]
name = "django-rtf"
version = "0.1.0"
description = ""
authors = ["Gabriel"]

[tool.poetry.dependencies]
python = "^3.12"
django = ">=6.0.3,<7.0.0"
djangorestframework = ">=3.17.1,<4.0.0"
psycopg2-binary = ">=2.9,<3.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"
pytest-django = ">=4.12.0,<5.0.0"

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py", "tests.py"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
